Increase maximum IR transmission length. Standardize and document endianness of transmissions

+++ LaserTagTransceiver.X/IRReceiver.h
@@ -10,9 +10,10 @@ void irReceiver_shutdown(void);
 void irReceiver_interruptHandler(void);
 
 // Returns true and copies the transmission data and length, in bits, into the out parameters if a transmission was
-// received since the last call to tryGetTransmissionData. Returns false otherwise. The contents of data_out are
+// received since the last call to tryGetTransmissionData. Returns false otherwise. data_out must point to a static
+// buffer of at least ((MAX_TRANSMISSION_LENGTH + 7) / 8 bytes. The contents of data_out are
 // undefined when this function returns false.
-bool irReceiver_tryGetTransmission(uint16_t* data_out, uint8_t* data_length_out);
+bool irReceiver_tryGetTransmission(uint8_t* data_out, uint8_t* data_length_out);
 
 #define DEBUG_RECEIVED_PULSE_LENGTHS
 #ifdef DEBUG_RECEIVED_PULSE_LENGTHS
